fake-news-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
huggingface-hub>=0.22.0
python-dotenv==1.0.0
transformers>=4.30.0
torch>=2.0.0
